requirements.txt
# AI Dubbing Tool Requirements
# Core dependencies for the dubbing tool

# Audio processing
pydub>=0.25.1
soundfile>=0.12.1
librosa>=0.10.0

# Machine learning and deep learning
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0

# Whisper dependencies
openai-whisper>=20231117
tiktoken>=0.5.0
numba>=0.58.0
more-itertools>=8.14.0

# F5-TTS dependencies
accelerate>=0.33.0
bitsandbytes>0.37.0
cached_path
click
datasets
ema_pytorch>=0.5.2
hydra-core>=1.3.0
jieba

omegaconf
pydantic<=2.10.6
pypinyin
safetensors
tomli
torchdiffeq
tqdm>=4.65.0
transformers
transformers_stream_generator
unidecode
vocos
wandb
x_transformers>=1.31.14

# Additional utilities
pathlib2>=2.3.7
typing-extensions>=4.0.0
psutil>=5.8.0
scipy>=1.9.0
moviepy>=1.0.3

# Web interface
gradio>=3.0.0 